Add Adventure audience to Instance

+++ src/main/java/net/minestom/server/network/packet/server/play/BlockChangePacket.java
@@ -9,7 +9,7 @@ import org.jetbrains.annotations.NotNull;
 public class BlockChangePacket implements ServerPacket {
 
     public BlockPosition blockPosition;
-    public int blockStateId;
+    public short blockStateId;
 
     @Override
     public void write(@NotNull BinaryWriter writer) {
